refactor(db): use json file rather than sql
Avoid over-engineering the solution and keep it simple where possible for convenience. Since the log
is just storing str-str key-value pairs. There's no enforcement that is required by something as
powerful as SQL. Therefore, a simple JSON flat file is more than enough. This does limit us to
essentially working in-memory but again, this project is not really meant for production so it does
not really matter.