feat: revert back to multiple json files
Database will actually be separate from log. We can readily push and pull from the database.
However, the log is a trail of our work to get the database to that point. Therefore, to save on
computational resources, it is best to use a dedicated database file lest we recompute the last
stored value for the key inside the log.