feat(db): add index field to entries
It would be impractical to rely on the JSON objects to provide the only form of validation for log
entries. During the AppendEntries RPC, the index is crucial to comparing and deleting future log
entries. Therefore, they need to be passed with the RPC messages. This can likely be optimized in
some low-level implementations.