refactor: remove AddServer and RemoveServer RPCs
These distract away from the essence of the Raft consensus algorithm and are merely extensions.
After some deliberation, I decided it was best to not have these in the repositories. The extensions
discussed in the Raft paper would be projects for students and enthusiasts to try to implement on
their own.